requirements.txt
fastapi
uvicorn
psycopg2-binary
werkzeug
pydantic
